Fix PR 33 regression closeout
