+++ package.json
@@ -28,6 +28,7 @@
     "validate:skill-refund": "node -e \"Promise.all([import('node:path'), import('node:url')]).then(([{ resolve }, { pathToFileURL }]) => import(pathToFileURL(resolve(process.env.INIT_CWD || process.cwd(), 'server/validate-skill-refund.js')).href))\"",
     "validate:open-progression": "node server/validate-open-progression-consistency.js",
     "validate:mage-progression": "node server/validate-mage-progression.js",
+    "validate:friendly-wolf-damage": "node server/validate-friendly-wolf-damage.js",
     "validate:dev-start": "node -e \"Promise.all([import('node:path'), import('node:url')]).then(([{ resolve }, { pathToFileURL }]) => import(pathToFileURL(resolve(process.env.INIT_CWD || process.cwd(), 'server/validate-dev-start.js')).href))\"",
     "validate:gameplay": "node -e \"Promise.all([import('node:path'), import('node:url')]).then(([{ resolve }, { pathToFileURL }]) => import(pathToFileURL(resolve(process.env.INIT_CWD || process.cwd(), 'server/run-validation-suite.js')).href).then((module) => module)).then(() => {})\" gameplay",
     "validate:loc": "node -e \"Promise.all([import('node:path'), import('node:url')]).then(([{ resolve }, { pathToFileURL }]) => import(pathToFileURL(resolve(process.env.INIT_CWD || process.cwd(), 'server/validate-loc.js')).href))\"",
